+++ regex/src/com.oracle.truffle.regex.test/src/com/oracle/truffle/regex/tregex/parser/flavors/PythonFlagsTest.java
@@ -60,6 +60,5 @@ public class PythonFlagsTest {
         assertTrue(parse("t").isTemplate());
         assertTrue(parse("u").isUnicodeExplicitlySet());
         assertTrue(parse("x").isVerbose());
-        assertTrue(parse("y").isSticky());
     }
 }