+++ regex/CHANGELOG.md
@@ -9,6 +9,7 @@ This changelog summarizes major changes between TRegex versions relevant to lang
 * Added `MustAdvance` and `PythonMode` options for better Python support.
 * Support for named character escapes in Python regular expressions.
 * Limited support for \G anchors in Ruby regular expressions.
+* Dropped sticky flag from Python regular expressions.
 
 ## Version 22.0.0